django_sheets/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
amqp==5.1.1
    # via kombu
asgiref==3.5.2
    # via django
async-timeout==4.0.2
    # via redis
backports-zoneinfo==0.2.1
    # via django
billiard==3.6.4.0
    # via celery
cachetools==5.1.0
    # via
    #   -r requirements.in
    #   google-auth
celery==5.2.6
    # via
    #   -r requirements.in
    #   flower
certifi==2022.5.18.1
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-compile-multi
    #   pip-tools
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
deprecated==1.2.13
    # via redis
django==4.0.4
    # via -r requirements.in
flake8==4.0.1
    # via -r requirements.in
flower==1.0.0
    # via -r requirements.in
google-api-core==2.8.0
    # via google-api-python-client
google-api-python-client==2.48.0
    # via -r requirements.in
google-auth==2.6.6
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
google-auth-httplib2==0.1.0
    # via
    #   -r requirements.in
    #   google-api-python-client
googleapis-common-protos==1.56.1
    # via google-api-core
gunicorn==20.1.0
    # via -r requirements.in
httplib2==0.20.4
    # via
    #   -r requirements.in
    #   google-api-python-client
    #   google-auth-httplib2
    #   oauth2client
humanize==4.1.0
    # via flower
idna==3.3
    # via requests
jinja2==3.1.2
    # via -r requirements.in
kombu==5.2.4
    # via celery
lxml==4.8.0
    # via -r requirements.in
markupsafe==2.1.1
    # via jinja2
mccabe==0.6.1
    # via flake8
oauth2client==4.1.3
    # via -r requirements.in
packaging==21.3
    # via redis
pep517==0.12.0
    # via pip-tools
pip-compile-multi==2.4.5
    # via -r requirements.in
pip-tools==6.6.2
    # via pip-compile-multi
prometheus-client==0.14.1
    # via flower
prompt-toolkit==3.0.29
    # via click-repl
protobuf==3.20.1
    # via
    #   google-api-core
    #   googleapis-common-protos
psycopg2==2.9.3
    # via -r requirements.in
psycopg2-binary==2.9.3
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   -r requirements.in
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via
    #   -r requirements.in
    #   google-auth
    #   oauth2client
pycodestyle==2.8.0
    # via
    #   -r requirements.in
    #   flake8
pyflakes==2.4.0
    # via
    #   -r requirements.in
    #   flake8
pyparsing==3.0.9
    # via
    #   httplib2
    #   packaging
python-dateutil==2.8.2
    # via -r requirements.in
python-dotenv==0.20.0
    # via -r requirements.in
pytz==2022.1
    # via
    #   celery
    #   flower
redis==4.3.1
    # via -r requirements.in
requests==2.27.1
    # via
    #   -r requirements.in
    #   google-api-core
rsa==4.8
    # via
    #   -r requirements.in
    #   google-auth
    #   oauth2client
six==1.16.0
    # via
    #   -r requirements.in
    #   click-repl
    #   google-auth
    #   google-auth-httplib2
    #   oauth2client
    #   python-dateutil
sqlparse==0.4.2
    # via
    #   -r requirements.in
    #   django
tomli==2.0.1
    # via pep517
toposort==1.7
    # via pip-compile-multi
tornado==6.1
    # via flower
uritemplate==4.1.1
    # via
    #   -r requirements.in
    #   google-api-python-client
urllib3==1.26.9
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
wheel==0.37.1
    # via pip-tools
wrapt==1.14.1
    # via deprecated
xmltodict==0.13.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
